[script] refactoring: split TypeChecker from resolve_types
